# Common.Contracts/openapi_v3.yaml
openapi: 3.0.4
info:
  title: LPshows.live API v3
  description: This is the API for the Linkin Park Concert Calendar fan project
  version: 3.0.0
servers:
  - url: /
  - url: http://localhost:5133
paths:
  /v3/Concerts:
    post:
      tags:
        - Concerts
      summary: Creates a new concert in the database
      operationId: CreateConcert
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConcertRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/CreateConcertRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/CreateConcertRequest"
        required: true
      responses:
        "200":
          description: the created concert
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: POST
        uri: https://api.dev.lpshows.felixsfd.de/v3/Concerts
      security:
        - Bearer: []
    get:
      tags:
        - Concerts
      summary: Returns a list of concerts that can be filtered
      operationId: GetConcerts
      parameters:
        - name: CountryCode
          in: query
          schema:
            type: string
        - name: Limit
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: string
            format: uint32
        - name: Skip
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: string
            format: uint32
        - name: OrderBy
          in: query
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: List of concerts including referenced objects
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConcertDetails"
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConcertDetails"
            text/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConcertDetails"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/Concerts
  /v3/Concerts/{concertId}:
    put:
      tags:
        - Concerts
      summary: Updates a concert in the database
      operationId: UpdateConcert
      parameters:
        - name: concertId
          in: path
          description: ID of the concert to update
          required: true
          schema:
            type: string
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConcertRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/UpdateConcertRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/UpdateConcertRequest"
        required: true
      responses:
        "200":
          description: no content
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: PUT
        uri: https://api.dev.lpshows.felixsfd.de/v3/Concerts/{concertId}
        requestParameters:
          integration.request.path.concertId: method.request.path.concertId
      security:
        - Bearer: []
    get:
      tags:
        - Concerts
      summary: Returns a concert without loading the referenced objects like the venue
        details
      operationId: GetRawConcertById
      parameters:
        - name: concertId
          in: path
          description: ""
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/RawConcert"
            application/json:
              schema:
                $ref: "#/components/schemas/RawConcert"
            text/json:
              schema:
                $ref: "#/components/schemas/RawConcert"
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/Concerts/{concertId}
        requestParameters:
          integration.request.path.concertId: method.request.path.concertId
      security:
        - Bearer: []
    delete:
      tags:
        - Concerts
      summary: Deletes a concert. Note that the data will not be fully removed from
        the database. Admins will still be able to see deleted concerts.
      operationId: DeleteConcertById
      parameters:
        - name: concertId
          in: path
          description: ID of the concert to delete
          required: true
          schema:
            type: string
      responses:
        "200":
          description: no content
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: DELETE
        uri: https://api.dev.lpshows.felixsfd.de/v3/Concerts/{concertId}
        requestParameters:
          integration.request.path.concertId: method.request.path.concertId
      security:
        - Bearer: []
  /v3/Concerts/{concertId}/details:
    get:
      tags:
        - Concerts
      summary: Returns all details about a concert. This includes information about
        the venue and general location.
      operationId: GetConcertById
      parameters:
        - name: concertId
          in: path
          description: ""
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ConcertDetails"
            application/json:
              schema:
                $ref: "#/components/schemas/ConcertDetails"
            text/json:
              schema:
                $ref: "#/components/schemas/ConcertDetails"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/Concerts/{concertId}/details
        requestParameters:
          integration.request.path.concertId: method.request.path.concertId
...
  /v3/ConcertTypes:
    post:
      tags:
        - ConcertTypes
      summary: Creates a new type of concert
      operationId: CreateConcertType
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConcertTypeRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/CreateConcertTypeRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/CreateConcertTypeRequest"
        required: true
      responses:
        "200":
          description: ""
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: POST
        uri: https://api.dev.lpshows.felixsfd.de/v3/ConcertTypes
      security:
        - Bearer: []
    get:
      tags:
        - ConcertTypes
      summary: Returns all available concert types
      operationId: GetConcertTypes
      responses:
        "200":
          description: information about the concert types
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ConcertType"
            application/json:
              schema:
                $ref: "#/components/schemas/ConcertType"
            text/json:
              schema:
                $ref: "#/components/schemas/ConcertType"
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/ConcertTypes
      security:
        - Bearer: []
  /v3/ConcertTypes/{concertTypeId}:
    get:
      tags:
        - ConcertTypes
      summary: Returns information about a concert type
      operationId: GetTypeById
      parameters:
        - name: concertTypeId
          in: path
          description: ID of the concert type
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
      responses:
        "200":
          description: information about the concert type
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ConcertType"
            application/json:
              schema:
                $ref: "#/components/schemas/ConcertType"
            text/json:
              schema:
                $ref: "#/components/schemas/ConcertType"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/ConcertTypes/{concertTypeId}
        requestParameters:
          integration.request.path.concertId: method.request.path.concertTypeId
    put:
      tags:
        - ConcertTypes
      summary: Updates information of a concert type
      operationId: UpdateType
      parameters:
        - name: concertTypeId
          in: path
          description: ID of the concert type
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConcertTypeRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/UpdateConcertTypeRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/UpdateConcertTypeRequest"
        required: true
      responses:
        "200":
          description: updated information about the concert type
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ConcertType"
            application/json:
              schema:
                $ref: "#/components/schemas/ConcertType"
            text/json:
              schema:
                $ref: "#/components/schemas/ConcertType"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: PUT
        uri: https://api.dev.lpshows.felixsfd.de/v3/ConcertTypes/{concertTypeId}
        requestParameters:
          integration.request.path.concertId: method.request.path.concertTypeId
...
components:
  schemas:
...
    CityWithCountry:
      required:
        - country
        - countryCode
        - name
        - nativeName
      type: object
      properties:
        country:
          $ref: "#/components/schemas/Country"
        state:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/State"
        countryCode:
          type: string
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        stateCode:
          type: string
        name:
          type: string
        nativeName:
          type: string
    ConcertDetails:
      required:
        - id
        - concertType
        - venue
      type: object
      properties:
        id:
          type: string
        concertType:
          $ref: "#/components/schemas/ConcertType"
        tour:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/Tour"
        tourLeg:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/TourLeg"
        customTitle:
          type: string
        venue:
          $ref: "#/components/schemas/VenueWithDetails"
        postedStartTime:
          type: string
          format: date-time
        mainStageTime:
          type: string
          format: date-time
        doorsTime:
          type: string
          format: date-time
        lpuEarlyEntryTime:
          type: string
          format: date-time
        lpuEarlyEntryConfirmed:
          type: boolean
        expectedSetDurationMinutes:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        scheduleImageFile:
          type: string
        deletedAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/ConcertStatusValue"
    ConcertStatusValue:
      type: string
      enum: 
        - Planned
        - Running
        - Past
        - Cancelled
    ConcertType:
      required:
        - name
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    Country:
      required:
        - isoCode
        - name
        - nativeName
      type: object
      properties:
        isoCode:
          maxLength: 3
          minLength: 3
          type: string
        name:
          type: string
        nativeName:
          type: string
...
    CreateConcertRequest:
      type: object
      properties:
        concertTypeId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        tourId:
          type: string
        tourLegId:
          type: string
        customTitle:
          type: string
        venueId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        postedStartTime:
          type: string
          format: date-time
        mainStageTime:
          type: string
          format: date-time
        doorsTime:
          type: string
          format: date-time
        lpuEarlyEntryTime:
          type: string
          format: date-time
        lpuEarlyEntryConfirmed:
          type: boolean
        expectedSetDurationMinutes:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        scheduleImageFile:
          type: string
        status:
          $ref: "#/components/schemas/ConcertStatusValue"
    CreateConcertTypeRequest:
      required:
        - name
      type: object
      properties:
        name:
          type: string
...
    PreviousVenueName:
      required:
        - venueId
        - id
        - name
        - usedFrom
      type: object
      properties:
        venueId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        name:
          type: string
        usedFrom:
          type: string
          format: date
        usedUntil:
          type: string
          format: date
    RawConcert:
      required:
        - id
      type: object
      properties:
        id:
          type: string
        concertTypeId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        tourId:
          type: string
        tourLegId:
          type: string
        customTitle:
          type: string
        venueId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        postedStartTime:
          type: string
          format: date-time
        mainStageTime:
          type: string
          format: date-time
        doorsTime:
          type: string
          format: date-time
        lpuEarlyEntryTime:
          type: string
          format: date-time
        lpuEarlyEntryConfirmed:
          type: boolean
        expectedSetDurationMinutes:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        scheduleImageFile:
          type: string
        deletedAt:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: "#/components/schemas/ConcertStatusValue"
    State:
      required:
        - countryCode
        - code
        - name
        - nativeName
      type: object
      properties:
        countryCode:
          type: string
        code:
          type: string
        name:
          type: string
        nativeName:
          type: string
...
    Tour:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        legs:
          type: array
          items:
            $ref: "#/components/schemas/TourLeg"
    TourLeg:
      required:
        - tourId
        - id
        - name
      type: object
      properties:
        tourId:
          type: string
        id:
          type: string
        name:
          type: string
...
    UpdateConcertRequest:
      type: object
      properties:
        concertTypeId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        tourId:
          type: string
          nullable: true
        tourLegId:
          type: string
          nullable: true
        customTitle:
          type: string
          nullable: true
        venueId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        postedStartTime:
          type: string
          format: date-time
        mainStageTime:
          type: string
          nullable: true
          format: date-time
        doorsTime:
          type: string
          nullable: true
          format: date-time
        lpuEarlyEntryTime:
          type: string
          nullable: true
          format: date-time
        lpuEarlyEntryConfirmed:
          type: boolean
        expectedSetDurationMinutes:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        scheduleImageFile:
          type: string
          nullable: true
        deletedAt:
          type: string
          nullable: true
          format: date-time
        status:
          $ref: "#/components/schemas/ConcertStatusValue"
    UpdateConcertTypeRequest:
      required:
        - name
      type: object
      properties:
        name:
          type: string
...
    VenueWithDetails:
      required:
        - venueNames
        - city
        - countryCode
        - currentName
        - timeZone
      type: object
      properties:
        venueNames:
          type: array
          items:
            $ref: "#/components/schemas/PreviousVenueName"
        city:
          $ref: "#/components/schemas/CityWithCountry"
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        countryCode:
          maxLength: 3
          minLength: 3
          type: string
        stateCode:
          maxLength: 3
          type: string
          nullable: true
        cityId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: string
          format: uint32
        currentName:
          type: string
        timeZone:
          type: string
        latitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type: string
          nullable: true
          format: double
        longitude:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type: string
          nullable: true
          format: double
    
    # Schemas from v2
...
  securitySchemes:
    Bearer:
      type: http
      description: JWT Bearer authentication
      scheme: Bearer
    ConcertsCognitoAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: cognito_user_pools
      x-amazon-apigateway-authorizer:
        type: cognito_user_pools
        providerARNs:
          - Fn::ImportValue:
              Fn::Sub: '${BaseProductName}-cognito-authorizer-user-pool'
...