# Server.Api/Controllers/ConcertsController.cs
using System;
using System.Linq;
using System.Threading;
using System.Threading.Tasks;
using Common.Contracts.Generated.Models;
using Common.Utils.Cache;
using LPCalendar.DataStructure.Tours;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.OutputCaching;
using Microsoft.Extensions.Logging;
using Server.Api.Auth;
using Server.Api.Cache;
using Service.Tours;
using CreateConcertRequestDto = LPCalendar.DataStructure.Tours.CreateConcertRequestDto;
using RawConcertDto = LPCalendar.DataStructure.Tours.RawConcertDto;
using UpdateConcertRequestDto = LPCalendar.DataStructure.Tours.UpdateConcertRequestDto;

namespace Server.Api.Controllers;

/// <summary>
/// Controller to manage concert data
/// </summary>
/// <param name="concertService"></param>
/// <param name="logger"></param>
[ApiController]
[Route("v3/[controller]")]
public class ConcertsController(ConcertService concertService, IOutputCacheStore outputCacheStore, ILogger<ConcertsController> logger) : ControllerBase
{
    /// <summary>
    /// Creates a new concert in the database
    /// </summary>
    /// <param name="request"></param>
    /// <returns>the created concert</returns>
    [HttpPost]
    [AuthorizeRoles]
    public async Task<CreatedAtActionResult> CreateConcert([FromBody] CreateConcertRequestDto request)
    {
        logger.LogDebug("Requested to create a new concert...");
        var concert = await concertService.CreateConcertAsync(request);
        logger.LogDebug("Created concert with id: {id}", concert.Id);
        await EvictConcertCacheAsync();
        return CreatedAtAction(nameof(GetRawConcertById), new { concertId = concert.Id }, concert);
    }
    
    /// <summary>
    /// Updates a concert in the database
    /// </summary>
    /// <param name="concertId">ID of the concert to update</param>
    /// <param name="request"></param>
    /// <returns>no content</returns>
    [HttpPut("{concertId}")]
    [AuthorizeRoles]
    public async Task<NoContentResult> UpdateConcert([FromBody] UpdateConcertRequestDto request, [FromRoute] string concertId)
    {
        logger.LogDebug("Requested to update the concert with id: {concertId}", concertId);
        var concert = await concertService.UpdateConcertAsync(concertId, request);
        logger.LogDebug("Updated concert with id: {id}", concert.Id);
        await EvictConcertCacheAsync();
        return NoContent();
    }

    /// <summary>
    /// Returns a concert without loading the referenced objects like the venue details
    /// </summary>
    /// <param name="concertId"></param>
    /// <returns></returns>
    [HttpGet("{concertId}")]
    [AuthorizeRoles]
    [OutputCache(PolicyName = CachePolicyNames.Short, Tags = [CacheTags.ConcertsAll])]
    public async Task<ActionResult<RawConcertDto>> GetRawConcertById([FromRoute] string concertId)
    {
        var concert = await concertService.GetConcertWithoutDetailsByIdAsync(concertId);
        return Ok(concert);
    }
    
    /// <summary>
    /// Returns all details about a concert. This includes information about the venue and general location.
    /// </summary>
    /// <param name="concertId"></param>
    /// <returns></returns>
    [HttpGet("{concertId}/details")]
    [OutputCache(PolicyName = CachePolicyNames.Medium, Tags = [CacheTags.ConcertsAll])]
    [CustomResponseCache(Duration = CacheExpiration.Default)]
    public async Task<ActionResult<ConcertDetailsDto>> GetConcertById([FromRoute] string concertId)
    {
        var concert = await concertService.GetConcertByIdAsync(concertId);
        return Ok(concert.ToDto());
    }

    /// <summary>
    /// Returns a list of concerts that can be filtered
    /// </summary>
    /// <param name="cancellationToken"></param>
    /// <param name="filter">Filter for the query</param>
    /// <returns>List of concerts including referenced objects</returns>
    [HttpGet]
    [CustomResponseCache(Duration = CacheExpiration.Default)]
    [OutputCache(PolicyName = CachePolicyNames.Medium, Tags = [CacheTags.ConcertsAll])]
    public async Task<ActionResult<ConcertDetailsDto[]>> GetConcertsAsync(CancellationToken cancellationToken, [FromQuery] GetConcertsFilterDto filter)
    {
        var concerts = await concertService.GetConcertsWithDetailsAsync(cancellationToken, filter)
            .Select(DtoMapper.ToDto)
            .ToArrayAsync(cancellationToken);
        logger.LogDebug("Retrieved {count} concert details.", concerts.Length);
        return Ok(concerts);
    }
    
    /// <summary>
    /// Deletes a concert. Note that the data will not be fully removed from the database. Admins will still be able to see deleted concerts.
    /// </summary>
    /// <param name="concertId">ID of the concert to delete</param>
    /// <returns>no content</returns>
    [HttpDelete("{concertId}")]
    [AuthorizeRoles]
    public async Task<NoContentResult> DeleteConcertById([FromRoute] string concertId)
    {
        await concertService.DeleteConcertAsync(concertId);
        await EvictConcertCacheAsync();
        return NoContent();
    }

    private async Task EvictConcertCacheAsync(CancellationToken cancellationToken = default)
    {
        try
        {
            await outputCacheStore.EvictByTagAsync(CacheTags.ConcertsAll, cancellationToken);
            logger.LogDebug("Evicted concerts cache.");
        }
        catch (Exception e)
        {
            logger.LogWarning(e, "Failed to evict concerts cache!");
        }
    }
}